django/requirements.txt
absl-py==2.2.2
alabaster==0.7.12
altair==4.1.0
anyio>=4.0.0,<5.0.0
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.5
Brotli==1.0.9
cachetools==5.5.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy>=1.0.0,<1.2.0
crossref==0.1.2
crossrefapi==1.6.0
cryptography==41.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.1
django-countries==7.6.1
django-extensions==4.1
django-organizations==2.5.0
django-simple-history==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docutils==0.18.1
dominate==2.9.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
feedparser==6.0.11
filelock==3.18.0
flake8==6.1.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
flatbuffers==25.2.10
fonttools==4.25.0
fqdn==1.5.1
freezegun==1.5.1
fsspec==2025.3.2
gast==0.6.0
gdown==5.2.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.71.0
h5py==3.13.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.31.4
idna==3.4
imagesize==1.4.1
iniconfig==1.1.1
ipykernel==6.28.0
ipython==8.20.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.5.0
json5==0.9.6
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
keras==2.15.0
kiwisolver==1.4.4
libclang==18.1.1
lightgbm>=4.0.0,<4.7.0
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.3.5
MarkupSafe==2.1.3
# Add platform-specific dependencies for matplotlib and numpy 
matplotlib>=3.7.0,<3.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
millify==0.1.1
mistune==2.0.4
ml-dtypes==0.2.0
mpmath==1.3.0
munkres==1.1.4
mypy==1.5.1
mypy-extensions==1.0.0
namex==0.0.9
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.4.2
notebook_shim==0.2.3
numexpr==2.8.7
numpy>=1.24.0,<1.27.0
oauthlib==3.2.2
ohio==0.5.0
opt_einsum==3.4.0
optree==0.15.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pep8==1.7.1
pexpect==4.9.0
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==4.25.7
psutil==5.9.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.21
pydub==0.25.1
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.9.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.16
PyQt5-sip==12.13.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
regex==2024.11.6
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.10.6
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
sgmllib3k==1.0.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
SpeechRecognition==3.10.4
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.30
sqlparse==0.5.3
stack-data==0.2.0
stopwords==1.0.1
sympy==1.14.0
tabulate==0.8.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
# Platform-specific TensorFlow dependencies
tensorflow>=2.11.0,<2.16.0; platform_system!="Darwin" or platform_machine!="arm64"
tensorflow-macos>=2.11.0,<2.16.0; platform_system=="Darwin" and platform_machine=="arm64"
tensorflow-metal>=0.7.0,<1.2.0; platform_system=="Darwin" and platform_machine=="arm64"
termcolor==3.1.0
terminado==0.17.1
tf-keras==2.15.0
threadpoolctl==3.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toolz==0.12.1
torch==2.7.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.7.1
transformers==4.38.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
visitor==0.1.3
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==3.0.3
win-inet-pton==1.1.0
wrapt==1.14.1
xhtml2pdf==0.2.2
